+++ refinedstorage-neoforge/src/main/java/com/refinedmods/refinedstorage/neoforge/ModInitializer.java
@@ -72,6 +72,7 @@ import com.refinedmods.refinedstorage.common.support.packet.s2c.AutocraftingPrev
 import com.refinedmods.refinedstorage.common.support.packet.s2c.AutocraftingResponsePacket;
 import com.refinedmods.refinedstorage.common.support.packet.s2c.AutocraftingTaskCompletedPacket;
 import com.refinedmods.refinedstorage.common.support.packet.s2c.EnergyInfoPacket;
+import com.refinedmods.refinedstorage.common.support.packet.s2c.ExportingIndicatorUpdatePacket;
 import com.refinedmods.refinedstorage.common.support.packet.s2c.GridActivePacket;
 import com.refinedmods.refinedstorage.common.support.packet.s2c.GridClearPacket;
 import com.refinedmods.refinedstorage.common.support.packet.s2c.GridUpdatePacket;
@@ -761,6 +762,11 @@ public class ModInitializer extends AbstractModInitializer {
             AutocraftingTaskCompletedPacket.PACKET_TYPE,
             AutocraftingTaskCompletedPacket.STREAM_CODEC,
             wrapHandler((packet, ctx) -> AutocraftingTaskCompletedPacket.handle(packet))
+        );
+        registrar.playToClient(
+            ExportingIndicatorUpdatePacket.PACKET_TYPE,
+            ExportingIndicatorUpdatePacket.STREAM_CODEC,
+            wrapHandler(ExportingIndicatorUpdatePacket::handle)
         );
     }