+++ refinedstorage-network-api/src/main/java/com/refinedmods/refinedstorage/api/network/autocrafting/AutocraftingNetworkComponent.java
@@ -4,13 +4,11 @@ import com.refinedmods.refinedstorage.api.autocrafting.Pattern;
 import com.refinedmods.refinedstorage.api.autocrafting.preview.PreviewProvider;
 import com.refinedmods.refinedstorage.api.autocrafting.status.TaskStatusProvider;
 import com.refinedmods.refinedstorage.api.autocrafting.task.ExternalPatternSinkProvider;
-import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
 import com.refinedmods.refinedstorage.api.network.NetworkComponent;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.api.storage.Actor;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import javax.annotation.Nullable;
 
@@ -34,5 +32,11 @@ public interface AutocraftingNetworkComponent
     @Nullable
     PatternProvider getProviderByPattern(Pattern pattern);
 
-    Optional<TaskId> ensureTask(ResourceKey resource, long amount, Actor actor);
+    EnsureResult ensureTask(ResourceKey resource, long amount, Actor actor);
+
+    enum EnsureResult {
+        MISSING_RESOURCES,
+        TASK_ALREADY_RUNNING,
+        TASK_CREATED
+    }
 }